requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
dj-database-url==3.1.2
Django==6.0.2
fonttools==4.61.1
gunicorn==25.1.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
paho-mqtt==2.1.0
pillow==12.1.1
psycopg2-binary==2.9.11
pyparsing==3.3.2
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
